requirements.txt
numpy<2
opencv-python==4.10.0.84
pyotp==2.9.0
qrcode==7.4.2
pyzbar==0.1.9
playwright
python-telegram-bot==20.7
apscheduler==3.10.4
bs4
lxml